Use 32-bit counter values

+++ src/gpu_ib/queue_pair.cpp
@@ -60,7 +60,7 @@ __device__ void QueuePair::quiet() {
 
   while (true) {
     bool done{false};
-    uint64_t quiet_amount{0};
+    uint32_t quiet_amount{0};
     uint32_t wave_cq_consumer{0};
     do {
       uint32_t posted = __hip_atomic_load(&quiet_posted, __ATOMIC_RELAXED, __HIP_MEMORY_SCOPE_AGENT);
@@ -69,7 +69,7 @@ __device__ void QueuePair::quiet() {
       if (!(posted - completed)) {
         return;
       }
-      uint64_t quiet_val = posted - active;
+      uint32_t quiet_val = posted - active;
       if (!quiet_val) {
         continue;
       }
@@ -83,8 +83,8 @@ __device__ void QueuePair::quiet() {
       done = __shfl(done, leader_phys_lane_id);
     } while (!done);
     wave_cq_consumer = __shfl(wave_cq_consumer, leader_phys_lane_id);
-    uint64_t my_cq_consumer = wave_cq_consumer + my_logical_lane_id;
-    uint64_t my_cq_index = my_cq_consumer % cq_cnt;
+    uint32_t my_cq_consumer = wave_cq_consumer + my_logical_lane_id;
+    uint32_t my_cq_index = my_cq_consumer % cq_cnt;
 
     if (my_logical_lane_id < quiet_amount) {
       volatile mlx5_cqe64 *cqe_entry = &cq_buf[my_cq_index];
@@ -112,7 +112,7 @@ __device__ void QueuePair::quiet() {
       __atomic_signal_fence(__ATOMIC_SEQ_CST);
     }
     if (is_leader) {
-      uint64_t posted {0};
+      uint32_t posted {0};
       do {
         posted = __hip_atomic_load(&quiet_completed, __ATOMIC_RELAXED, __HIP_MEMORY_SCOPE_AGENT);
       } while (posted != wave_cq_consumer);
@@ -134,21 +134,21 @@ __device__ void QueuePair::post_wqe_rma(int pe, int32_t size, uintptr_t *laddr, 
   bool is_leader{my_logical_lane_id == 0};
   const uint64_t leader_phys_lane_id = __ffsll((unsigned long long)activemask) - 1;
   uint8_t num_wqes{num_active_lanes};
-  uint64_t wave_sq_counter{0};
+  uint32_t wave_sq_counter{0};
 
   if (is_leader) {
     wave_sq_counter = __hip_atomic_fetch_add(&sq_posted, num_wqes, __ATOMIC_RELAXED, __HIP_MEMORY_SCOPE_AGENT);
   }
   wave_sq_counter = __shfl(wave_sq_counter, leader_phys_lane_id);
-  uint64_t my_sq_counter = wave_sq_counter + my_logical_lane_id;
-  uint64_t my_sq_index = my_sq_counter % sq_wqe_cnt;
+  uint32_t my_sq_counter = wave_sq_counter + my_logical_lane_id;
+  uint32_t my_sq_index = my_sq_counter % sq_wqe_cnt;
 
   do {
-    uint64_t posted = __hip_atomic_load(&sq_db_touched, __ATOMIC_RELAXED, __HIP_MEMORY_SCOPE_AGENT);
-    uint64_t sunk = __hip_atomic_load(&sq_sunk, __ATOMIC_RELAXED, __HIP_MEMORY_SCOPE_AGENT);
-    uint64_t num_active_sq_entries = posted - sunk;
-    uint64_t num_free_entries = min(sq_wqe_cnt, cq_cnt) - num_active_sq_entries;
-    uint64_t num_entries_until_wave_last_entry = wave_sq_counter + num_active_lanes - posted;
+    uint32_t posted = __hip_atomic_load(&sq_db_touched, __ATOMIC_RELAXED, __HIP_MEMORY_SCOPE_AGENT);
+    uint32_t sunk = __hip_atomic_load(&sq_sunk, __ATOMIC_RELAXED, __HIP_MEMORY_SCOPE_AGENT);
+    uint32_t num_active_sq_entries = posted - sunk;
+    uint32_t num_free_entries = min(sq_wqe_cnt, cq_cnt) - num_active_sq_entries;
+    uint32_t num_entries_until_wave_last_entry = wave_sq_counter + num_active_lanes - posted;
     if (num_free_entries > num_entries_until_wave_last_entry) {
       break;
     }
@@ -164,7 +164,7 @@ __device__ void QueuePair::post_wqe_rma(int pe, int32_t size, uintptr_t *laddr, 
   __atomic_signal_fence(__ATOMIC_SEQ_CST);
 
   if (is_leader) {
-    uint64_t posted {0};
+    uint32_t posted {0};
     do {
       posted = __hip_atomic_load(&sq_db_touched, __ATOMIC_RELAXED, __HIP_MEMORY_SCOPE_AGENT);
     } while (posted != wave_sq_counter);
